fix: add auto-update
